native-language-identification/requirements.txt
# Native Language Identification Requirements

# Core Dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.11.0
audioread>=3.0.0

# Machine Learning
scikit-learn>=1.0.0
datasets>=2.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
huggingface-hub>=0.14.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=5.0.0
ipython>=8.0.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0
ipywidgets>=8.0.0

# Optional: For faster computation
# numba>=0.56.0
# cupy-cuda11x>=11.0.0  # For CUDA 11.x
